feat:Permissions user

+++ src/modules/users/users.service.ts
@@ -2,6 +2,7 @@ import {
   ConflictException,
   Injectable,
   NotFoundException,
+  ForbiddenException,
 } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
@@ -43,21 +44,39 @@ export class UsersService {
     return findUser;
   }
 
-  async update(id: string, updateUserDto: UpdateUserDto) {
+  async update(id: string, updateUserDto: UpdateUserDto, userId: string) {
     const findUser = await this.usersRepository.findOne(id);
+
+    const findEmailUser = await this.usersRepository.findByEmail(
+      updateUserDto.email,
+    );
+
+    if (id !== userId) {
+      throw new ForbiddenException('Insufficient permission');
+    }
 
     if (!findUser) {
       throw new NotFoundException('User not found!');
     }
+
+    if (findEmailUser) {
+      throw new ConflictException('Email already exists');
+    }
+
     return this.usersRepository.update(id, updateUserDto);
   }
 
-  async remove(id: string) {
+  async remove(id: string, userId: string) {
     const findUser = await this.usersRepository.findOne(id);
 
     if (!findUser) {
       throw new NotFoundException('User not found!');
     }
+
+    if (id !== userId) {
+      throw new ForbiddenException('Insufficient permission');
+    }
+
     return this.usersRepository.delete(id);
   }
 }
